simplify and move reducers to expr/agg (#1893)
This commit simplifies reducers and group-by so that all aggregators are
now decomposable.  It also clarifies the new agg.Function interface
with respect to partials.  Since first() and last() are not decomposable,
we removed them in favor of an any() aggregator as in systems like
clickhouse.  If and when we re-introduced non-decomposable aggregators,
we should call them something else and implement them differently at
the group-by layer since they cannot be decomposed the same way for
distributed aggregation (e.g., they need to work more like join/sort
depending on their sensitivity to input order).

We also reduced the footprint of each group-by row by removing the stats
counters that weren't used and moving "expr" and "where" interface pointers
out of group-by row and into the group-by template for each row since
these pointers are the same for every row in the table.

These changes allow us to use aggregators now in expression context,
which will be useful in coming PRs.  It also sets us up to do a further
refactor in group-by to better handle distributed group-bys and merging
of partials across a cluster.