Fix Zeek record unflattening for field with record prefix
Zeek record unflattening incorrectly considers a field to be part of a
nested record if the nested record name is a prefix of the field name,
causing zio/zeekio.Reader to include bound_p in the bound record and
request_p in the request record when reading a SOCKS log.  Fix this in
zio/zeekio.unflattenRecord.