fix flattener for nested nulls
This commit fixes a bug in the flattener where nested nulls were
not properly output for every column in the flattened record
(i.e., a null value for an embedded record did not properly recurse
into the nested record so only one null value was output instead
of a null for each flattened value).

While fixing this we added zcode.AppendNull() to clarify that the API
for appending null is the same for primitives and containers.