Replace zbuf.ReadBatch with zbuf.NewPuller
zbuf.ReadBatch can signal EOS in two ways: by returning a nil batch and
nil error or a short batch and nil error.  This undocumented behavior
leads to a bug in zbuf.Scanner.Pull, which does not check for a short
batch.  Fix the bug by replacing ReadBatch with zbuf.NewPuller.