make fuse stable
This commit changes the fuser so the output type columns of
the union schema are stable.  Previously, the column order
was dependent on the type IDs, which could be non-deterministic
when scanning paralllel inputs.  The new code determines the
column order from the order of the types encountered in the
incoming stream.