fix rename bug
This commit fixes a bug in the rename processor where new records
would be created with the wrong alias type.  This will all be fixed
when we rework the zng.Record data structure and batch, but for now
we simplify the code here to always create a new zng.Record so the
proper alias pointer is created.