fix bug translating primitive-type aliases between type contexts
This commit fixes a bug where an alias for a primitive type was not
being translated into the target context and the alien alias type
instead made its way into the context.  This manifested with the
tzng writer emitting multiple alias typedefs for the same type,
then the tzng reader complained.