fix rename bug (#2046)
This commit fixes a bug in the rename processor where new records
would be created with the wrong zng.Type "Alias" pointer in the 
zng.Record.  This will all be fixed when we rework the zng.Record
data structure and batch, but for now we simplify the code here to always
create a new zng.Record so the proper alias pointer is created.