fix scanning of zson primitives (#2058)
The logic to scan zson primitive values did not properly handle the
recent changes to read input as a stream and had presumed all of
the needed input was in the cursor.  This commit fixes peekPrimitive()
to make sure there is data up to the next whitespace, comma, or EOF
(since none of this can appear in a primitive value).