first step toward unifying search and expressions (#2072)
This commit makes changes toward unifying search and expressions.
There is still much to be done, but this is a first step forward.
As part of this, we introduced a generator pattern as the low-level
construct for iterator wildcard field comparisons (e.g., "*=foo").
The first cut uses a "select" function to select and generate
sequences of values.  The syntax is a proposal to server as a
starting point for discussion.

We also eliminated "**" for now and the "=~" in favor of "in" for
CIDR matches.

There is now a scope abstraction in the compiler which sets us
up for named constants, variables, etc. as well as handling the
"$" reference for generated values.

Any aggregator can now be applied to a generated sequence.

Generated sequences can be modified or filtered with methods.
Currently there is support for just map() and filter().

The grammar has been refactored quite a bit in making this changes
and should be easier to extend and maintain.  There is one more big
move we want to make in unifying the standard-expression and
search-expression syntaxes.

In the runtime, the filter predicates are now all optimizations
of binary expressions and ahve been moved to package expr.

The where clause in group-by is now a search expression.

The match() function can be used in expression context to escape
the expression syntax into the search syntax.  Match() is boolean
valued.

The ast.BinaryExpr nodes now use lower case "and" and "or" for
the binary operator field.  This makes everything consistent now
and fixes a bug where case mismatch could occur in certain
situations.

The notes in zql/expr.md are rough and should be eventally
replaced with user docs and/or a design doc.