fix join bug with memory management
The join proc did not correctly do a zng.Record.Keep() while building
the join set array causing the underlying batch to be freed while
record references still existed.  The fix is to call keep in the right
place.  Also, the joinkey memory could also be freed while in use
so it's now copied to a buffer in the join proc struct.