cleanup zson boomerang tests and fix zson bugs
This commit removes the clunky test files from zson/ztests/auto and replaces
with them with logic to compute this test files on the fly.  This required
some refactoring of package ztest.

With this change, new zson bugs were found in the time since the zson auto
tests were created.  These bugs and fixes include:

 o  "0x" with no byte values following is now an empty bytes slice
 o +Inf,-Inf,Nan etc is parsed correctly by zson
 o unsigned integers larger than MaxInt64 are handled properly by zson
 o there is a Go bug with duration of MinInt64 and this is now worked around