add "function proc", exists(), and fields() and drop parens from types (#2131)
This commit adds the notion of a "function proc" which is a FunctionCall
ast node appearing as a proc.  This is used to represent the ambiguity
between a simple reducer like "count()" used without an assignment and
a boolean function like "is('{s:string}')".  The ambiguity is resolved by
the compiler by looking up the name and choosing either a GroupByProc
or a FilterProc to form the appropriate representation.  To do this,
there is now a SemanticTransform() step in package compiler.  This replaces
compiler.ReplaceGroupByProcDurationWithKey() and does this key transform
along with any other semantic transformation we need to do post parsing.

We also added an option to expression functions to ask that the root
record be passed to them.  This is used by "is()" to do type matching
on "." or an explicit second argument.

We also added the is(), fields(), and exists() functions.  is()
provides type comparisons and fields() returns an array of dotted-field
strings representing the fields in a type value (that is a record type)
or a record value.  This latter function is useful during data shape
exploration to layout the fields encountered as arrays strings and
cut down on the noise.

The exists() function returns true iff all of it's arguments evaluate
to a non-missing, non-error value.  e.g., exists(foo) is true if the
field foo exists and exists(foo / 0) is always false.  exists(a.b,a.c)
is true if both fields exist unless their value is an error value.