add zson to auto-detector (#2123)
This commit adds zson to the auto detector and arranges for zson
to parse ndjson when auto-detected, unless a legacy ndjson type
config is provded, in which case the ndjson parser is used.
This allows us to support legacy type configs while we update
the user base with the new Z-shaper approach to ingesting json.

One ramification is that the zson parser converts numbers that
are textually an integer to int64 instead of json's float64.
We plan to add a json-strict mode to the zson parser, which
would allow us to circumvent this though this would be an explicit
flag and not inferred by default in the autodetection path.
See issue #2122.

Another ramification is that the ndjson parser assume null:string
for unknown nulls whereas zson assumes null:null.  Now that we have
shaping and type algebra, this later approach is prefered because
the shaper can know from null:null that a type is indeterminate
compared to null:string, where the type was arbitrarily chosen
to be string.