clean up conventions for missing values (#2154)
This commit creates zng.ErrMissing and zng.Missing to work toward
a unified way of handling missing values (i.e., expressions that
refer to fields that don't exist).  ErrMissing is a Go error
that can be returned up the chain to signal this condition while
Missing is a value that can be returned in zng data when a
result calls for embedding a Missing value instead of ignoring
it all together.