add flow DSL package
This commit adds the flow DSL abstraction to the zq system.
Now the PEG parser generates an AST, which is semantically
anaylized and turned into a flow.Graph DSL.  The flow.Graph
defines a Z job running on the system and is the API endpoint
for workers.  The client SearchRequest now uses Z source text
in its API and the parser AST, while still used by the compiler
subsystem, no longer appears in any API.