lake: Parallelize tsdir fetch
The walk of lake stored in s3 is slowed greated in that tsDirs are fetched
sequentially. The main thread sits by idly while the metadata for all chunks in
a tsdir are retrieved over the network. Speed this process up by running the
tsdir fetches each in a separate goroutine.

Because it is important that results be returned in order, save the retrieved
tsdirs in a min/max heap, and return the next min and max when it is received.

Closes #2017