simplify ast unpacker
This commit simplifies the ast unpacker by adding a generic package
for umarshaling interface types by keying on a configurable field
in the json.  The value of said key indicates which Go struct type
to unmarshal into via its type name (though this can be overriden
as evidence in our current ast where the reflector AddAs() method
is used).  Also, this requires that the Op field be filled in for
anything marshaled, which is get habit, because forthcoming code
that transforms ASTs and sends the transformed AST over the network
to workers must have all the Op fields present.