change -z to -I and allow multiple files
This commit changes the flag (-z) for running zq with the Z query in
a file.  The flag is replaced with -I (for "include") and the behavior
is a bit different: -I can be used multiple times and it is combined
with the normal Z query on the command line.  These combinations can
create some confusing syntax errors so the Z text parsed is display
when a syntax error occurs.