extract tzng-related out of packange zng and into tzngio
This commit does a surgical extraction of the TZNG-related
logic from the package zng and preares us for the unbundling
of the zeek format.  Unfortunately, we're leaving here a bit of
a mess because comprehensive Z literals have not yet been
implemented; as soon as this goes in, the mess can be removed.
I've tried to leave comments in pointing to the messy bits
that will soon going away so please don't spend much time
reviewing this on the PR.

These changes also nicely highlights the vestigal dependencies
on tzngio so we can now easily grep the source tree to see which
pieces still depend on tzngio and thus still need to be ported
the rest of the way to Z.

Finally, these changes are mostly all code movement, except we
changed the format of durations in NDJSON to conform to RFC 3339.