move around type system to reflect zson design
This commit provides a long-needed refacting to realign with the
zson/zng focus of the project.  The tzng type parser was moved
to zio/tzngio and the type context logic moved to package zson.
We wrapped the new zson.TypeContext in resolver.Context with the
idea of moving package resovler in a subsequent commit,
With resolver depending on zson, the marshaling needed to move
into zson.  There is also a wrapper for zng.Marshaler that points
back at zson.ZNGMarshaler.

At this point, the types in the type context table are all native
zson and the tzng types are isolated to package tzngio.

This is also a first step toward getting the tzng formatting and
parsing logic out of package zng.