fix bug where random data looked like non-record zng value
This commit fixes a bug where bad data that happens to parse
as zng caused a panic when the type ID of the zng value was
was not a record type.  The fix is to check the that top-level
values are records and return an error if they don't, which
otherwise allows the auto-detector to complete its task
and report an error.

Fixes #2205