simplify ast unpacker (#2188)
This commit simplifies the ast unpacker by adding a generic package
for umarshaling interface types by keying on a configurable field
in the json.  The value of said key indicates which Go struct type
to unmarshal into via its type name though this binding can be overridden
with using an explicit value in the unpack struct tag.