Add open telemetry tracing
First pass of adding open telemetry tracing instrumentation for zqd. In the
zservice helm chart, if is value otel.enabled is set to true, the AWS Distro
for OpenTelemetry's collector is deployed as a sidecar to each pod running zqd.
By default the AWS variant of the open telemetry collector sends traces to AWS
X-Ray.

TODO:

- [ ] Separate different zservice installs by group in X-Ray console.

- [ ] Change the requestID logger to use the open telemetry traceID.

- [ ] Add zqd configuration to setup global tracer provider. Enable this in
  helm chart when otel.enabled is set to true.

- [ ] Fix URL issue with the recruiter long poll trace.

- [ ] Add traces for compaction.

Closes #2143