add csv reader
This commit adds a reader for CSV formatted input.  It uses the
Go csv package, which conforms to RFC 4180.  CSV is not yet tied
into the auto-detector as it is currently too forgiving of a format
and accepts inputs that are clearly not intended as CSV.  See
issue #2315 for more info.

Closes #999.