fix time and duration casts
This commit fixes the time and duration casting logic to be more
flexible and reasonable. Strings in RFC-3339 can now be cast to
times or durations.  Numerics (float and int) are now presumed to
both be seconds and float64 rounding errors are allowed through
as there are better ways to avoid such rounding errors by using
ISO date/times or native ZNG times from a client.

We also changed the casting logic and iso() function to return
embedded Z errors rather than fatal Go errors to allow computation
to continue in the face of errors.

Given this new behavior and the upcoming additional of ZSON-oriented
Z literals, there is no longer a need for functions sec(), msec(),
etc so they have been removed.