Fix flaky test for #2293
The ztest for fixedbugs/2293 should have used the "-q" flag since the test does
not care about the output of "zar index create".

Never the less, the flaky test exposed an race conditition with the progress
update functionality, where the program would exit before the update logic
would have a chance to print all updates. Fix this as well.

Also the test was named after the wrong issue, update the name to reflect the
right issue.

Closes #2360