zbuf.NewCompareFn: Respect nullsMax on reverse
This came about because the merge proc on a parallel count by every on
descending sorted data was not handling null values in the same manner as the
child group by procs (i.e. For descending ordered data nil values should be the
last elements in a sorted list). Fix zbuf.NewCompareFn so that it sets nullsMax
to false if the input is in descending order.

Closes #2259