finalize zson sytnax
This commit changes the definition of a zson duration to be more
comprehensive.  We looked at RFC-3339 but the syntax is unergonomic
and does not cover sub-second granularity.  We decided to follow
the lead of Prometheus duration syntax but extend it with microseconds
and nanoseconds.

To implement this change, we created a nano.Duration type in
package nano.  Further, we changed nano.Span to use Duration instead
of int64.  This means json marshaled nano.Spans (e.g., in the zqd
service API) now use zson duration strings.