allow typedefs to be redefined
The commit relaxes the constraint that type names cannot be redefined
to a different type.  This follows the Z design principle of dealing
with whatever the user or client throws at the system as best we can.

The problem of schema versioning is outside the scope of the ZSON
format and can be solved at a higher layer with the appropriate Z
building blocks.

We also removed the end-of-sequence concept from the ZSON spec as
it is not relevant here as the definitions of type contexts and
redefinable types allows an implementation to create a seekable output.

Closes #1568