add missing(), has(), and nameof() (#2393)
This commit adds some sorely needed functions that will be useful for
data discovery, exploration, classification, and shaping.  missing()
returns true iff any of the arguments are missing.  has() returns
true iff all of the arguments are not missing.  nameof() returns
the type name of the single expression if the type is an alias;
otherwise, it returns missing.