add ISO time literal
This commit adds support to the Z grammar for ISO time literals
and exposes durations and time literals into the expression
grammar.  We also added support for searches in the generic
expression compilation path as this can occur when match() is
used in expression context, e.g., put b=match(foo), which
we used for the time/duration literal testing.