fix locking bug in zson context
This commit fixes a bug where unlock was called on an unlocked lock.

This was triggered when a typedef name was accessed that did
not yet exist.  We added a test to exercise this path with an
issue() function that looks for a type that is not defuned until
the second record in the input.