Add zbuf.Reader log post method
Add a method for client.Connection that allows users to pass a stream of
records to be ingested into zqd. These changes are needed for the load command
in the upcoming brimcap repo.

Add zngio.IOReader that uses a zngio.Writer to transform a zbuf.Reader into an
io.Reader.