Add ToStdio method to emitters
Have emitter.NewDir and emitter.NewFile return a new Emitter interface, which
is a zbuf.WriteCloser along with a function that returns true if the emitter is
writing to stdio (i.e. stdout or stderr). This can be used to write command
line functionality that changes depending on where zng records are being output
to.