add zson.FormatType() (#2511)
This commits add the FormatType() function that formats up a
zng.Type in its canonical form.  The ZSON() method on zng.Type
didn't quite work right because it would output multiple
embedded typedefs when only the first is needed.  We may want
to get rid of the ZSON() method at some point as is commented
in issue #2473.