handle Zed bytes values in marshal/unmarshal
This commit changes the marshaling code to output a Zed bytes type
for Go byte slices and byte arrays and to detect when a bytes type
is being unmarshaled into a byte slice or array and unmarshal
accordingly.

We also fixed a bug where UnarshalRecord failed when the out record
had a type name instead of a zng.TypeRecord.