split AST into a DAG piece and a parser piece
This commit separates the AST into a execution-oriented piece
that describes the computational flowgraph (DAG) and a parser-oriented
piece that represents the output of the Zed language grammar.

The optimization logic and query planner operates exlusively on
the DAG and workers pass DAG fragments around to execute a
distributed query.

The AST should not appear in any API and we should evolve
the outward facing search API to use either a text Zed language
description of the query or a DAG description.

At some point, we will have a parser that takes a canonical form
DAG language input and produces the DAG data structure, which
will be handy for test and debug.