improve ergonomics of zed lake commands (#2577)
Now that the dust is settling on the zed lake command set, this
commit unifies the command sets and arranges for all of the
lake-introspection commands (ls, log, status) to produce output
in ZNG format that can then be pretty-printed with "-f lake.
This change is important for a few reasons.  First, it makes
it easy to view introspection results any way you want with zq.
Second, it means that the service API can use exactly the same
lake interface as the CLI tools and there is no need to create
a special API messages for every type of introspection since
it's all just ZNG.  If the API allows for a format (content-encoding?),
then we can be compatible with JSON API's by simply using "-f json".
Third, it is the right abstraction for handling large results from
introspection as the are simply streamed as a result instead
of being packaged into a bulky array that potentially does not
fit into memory.

This commit also adds some fit and finish like terminal colors
(now for the commit log, but we should add to -Z and other places)
a username heuristic from commit messages, and a time-travel "-at"
flag for zed lake ls.

We also commented out much of the worker logic in package driver
as it's awkward dependencies caused an import cycle with the addition
of lakeio.  This prototype anyway does not have the right abstraction
flexibility for an optimizing execution planner and it will be
reworked soon.

Finally, we hit the known-but-undocumented bug with marshaling
mixed-type arrays and filed issue #2575 though we worked around
it by changing the format of the staged-commit scan.