colorize zson (#2621)
This commit add colorization of zson -Z text.  This is a simple start
that we can improve upon over time.  We should probably unify the
colors with the app (though the terminal color codes are limited and
aren't generic RGB values).  Type decorators are toned down in a
light gray to make them less obtrusive.  We also added -color to
outputflags so color can be disabled with -color=false and the
upcoming defaults settings