change top-level "." to "this"
Gordon pointed out that "." is weird and cryptic in expressions
like "any(.) by typeof(.)" and suggested "this" as in C++.
Just because jq has ".", jquery has "$", and underscore has "_"
doesn't mean we need to jump off the cliff too.

This commit adds "this" and leaves "." in the grammar for now
to see if we like it as a short cut for "this".  Now we can
say "any(this) by typeof(this)".  We can deprecate "." later
depending on how it goes.