rework dataflow model and compiler optimizations (#2669)
This commit is a large change to improve the way dataflow
is structured.  This change allows the optimizer to better
interact with the lake's "scan scheduler".  It also removes
the mess that ended up in package driver over time.  The driver
package is now restored to its roots, where it was meant to be
a simply interface over which to run a search query (e.g., from
the CLI, from the search endpoint, from the ingest shaper, etc).
There is no more mux.

Data source objects are now embedded in the AST and DAG as
"sources" and a From operator provides a mean to connect
sources with trunks of the input path.  Importantly, each trunk
has a convenient place to store pushdown logic, which will be
naturally carried over the network to distributed works via
the forthcoming run-DAG API.

The searchflags package was removed as from/to key ranges and
the data pool our now embedded in the Zed from operator.

To support different storage layers, there is now a proc.DataAdaptor
interfaces and implementations for the file system, the Zed lake,
and internal readers.

The "order" abstraction has been unified into a top-level package
called "order" and zbuf.Order has been deprecated.  This package
supports a simple order concept (asc/desc) plus a scan-direction
conception (asc/desc/unknown).  There is also an order.Layout
absaction to represent sort keys (primary, secondary, ...) and
an order with a text syntax "key1,key2,...:asc|desc".  The
pool create command uses this syntax with a new "-orderby" option.

The dataflow analysis in the compiler has been improved and
generalized.  It should be much easier to extend and add operators.

A stub for HTTP sources is now in place so it should be trivial
to add support from "from https://example.com/etc"

These changes made it trivial to support cross-pool queries.
A new test has been added to exercise a simple cross-pool join.