fix whitespace bug in ztest regexps
Many times I have written a regexp ztest rule only to be mystified
for why it didn't work.  The mystery lies in a yaml newline that
is embedded in the regexp requiring a silent match of a newline
immediately at the end of any regexp.  This commit adds a
string.TrimSpace to ztest whitespace to fix this.  This means
that you can't rely on whitespace at the start or end of a regexp
and must explicitly put any such whitespace in parens.  None of the
existing tests work this way.

We also removed an extraneous single quote from the regexp error
message.