rework storage engine abstraction (#2688)
The previous "iosrc" package provided a UFS-like interface whereas
the Zed system should be based on a cloud object model with a
UFS-emulation layer.  This commit converts iosrc to this new
interface abstraction and renames it"pkg/storage".

We also moved the immutable object caching implementation from
server/immcache to pkg/storage/cache and update the caches to
wrap and implement a storage.Engine.

We also changed storage.URI from struct{*url.URL} to url.URL
and used pointers throughout to be consistent with url.URL
usage patterns.  The stdio URI pattern has been changed to
elide the unnecessary "///" characters, e.g., stdio:stdin
instead of stdio:///stdin, which conforms with the URL RFC.

This change also exposed some structural problems with assumptions
when local files can be accessed.  For example, tests were passing
for s3 when using WriteIfNotExists but that should have failed.
There is now an issue for implementing PutIfNotExists on top of
S3 strong consistency and this is documented in the journal commit
code path.