add support for marshaling zng.Value fields (#2693)
This commit adds support for marshaling and unmarshaling zng Values
that are embedded in Go structs.  This allows us to encode arbitrary
zng.Values as part of native Go data structures, e.g., to encode
the upper/lower range boundary as zng.Values in the metadata struct
of an arbitrary Zed lake data object.

While here, we also improved the error message when trying to decode
a non-record Zed value into a Go struct.