add support for arbitrary pool keys
Much of the backend presumed the pool key was of type time.
This commit generalizes the scanning logic to allow pool keys
of any type.

The seek index has been simplified to use a simple, flat
index that is no long a "micro index".  Instead it is a simple,
single-level list of keys that is used at open to get the
scan range for a row object when the scan is smaller than the
whole object.  This design presumes the index will be cached
when running multiple queries over a sub-range within a row object.

While updating tests, we noticed the segment size and row_size
fields were virtually the same so we deleted the size field.

As part of this commit, the index package was updated to use
zson format for key parsing instead of the deprecated tzng format,
so we removed zio/tzngio/builder.go.