simplify seek index writer and CLI arg
This commit simplifies the seek index writer by moving the
logic for auto-stream-termination out of zngio and into
segment.Writer.  We also deleted the unused indexing code
in zngio along with the "zed index seek" command and related
support.

As previously agreed among the team, the streamax argument
to zed lake add and zed lake load was changed to seekstride
and the units changed from records to bytes.

Closes #2542