First cut of lake endpoints
The pr is a first cut of reorganizing the zed lake service api to mimic
the commands seen in zed lake. It also rearranges the client library so
it is lower level and has no knowledge (with the exception of errors) of
the data returned in api responses.

Part of #2651