fix dataflow through cut operators
This commit fixes a problem where a query would be incorrectly
parallelized by merging on the wrong key because sufficient
dataflow info was not properly tracked.  While we fixed this
we broke another optimization that doesn't seem important to
maintain (and anyway was based on flawed logic).  These tests
are marked "skip".  When we implement proper dataflow analysis
as part of issue #2756, we can turn the tests back on.