fix read/write race in journal.Queue
When reading HEAD, the journal can read the contents while
a writer has opened and truncated the contents but before the
writer has written the updated HEAD.  This happens only on
file systems, not cloud objects, as cloud writes are atomic.

The fix is to simply retry the read when the read is successful
but the content cannot be parsed.  This fix is preferred over
a more complicated locking model as the HEAD read is just advisory
and not required for correctness of the log abstraction.

Closes #2699