disable cut warnings for parallel lake queries (#2764)
When a query having a cut is parallelized, the warnings can occur
even though a field is found on one of the parallelized branches.
In fact, every branch that does not find the cut fields, issues a
warning.  This commit disables cut warning for queries that are
parallelized to avoid this problem.  A better approach is to warn
at compile-time based on the presence of fields gleened from type
information that will be stored with each segment.  This approach
will be implemented in a future PR.