fix bug in lakeio formatSegment() (#2816)
This commit fixes a simple bug in formatSegment() in lakeio to print the key range using zson.String() instead of zng.Value.String(). While we were here, we noticed system tests failing that relied upon the ADD order of objects in a commit log. Since squash uses snapshot which uses a Go map to figure out all the ADDs for a given snapshot, order is not (and need not be) preserved. We updated these tests to use zng output with some zq magic to handle the potential reordering.

The null weirdness will be fixed in #2818

closes #2803
closes #2817