encode type values in binary and move type context into zng (#2826)
This commit reworks the type value specification and implementation
so that ZNG-serialized type values have their own encoding independent
of the ZSON format.  This breaks the dependency between zng and zson
and will allow the zng package to be moved to the top-level zed package.

This also introduces a backward-incompatible change to the ZNG format as
type values are now encoded differently.  However, this change should
be minimally impactful as type values appear more often dynamically as
the result of a query and not so much in origin data.