zbuf.Merger: Emit EOS after run has completed
Package driver assumes that when an EOS is emitted from a puller the
flowgraph is no longer consuming the resources of proc.Context. In
zbuf.Merge the run goroutines are orphaned from EOF and would sometimes
call Warn on the proc Context even though Warn had been closed. Fix this
race.

Closes #2822