improve parse error formatting
When a parse error occurs, Zed tools print the full source code with no
source file information.  This is inhospitable when programs are long,
composed from files and the command line, or composed from multiple
files.  Improve this by printing only the source code line containing
the error along with a file name if appropriate.